feat: solve data-attribute and remove btn
Just needed to use indexOf the book itself in the createBookCard
function.

For the displayBook function, wondering if I can make it so that upon
clicking removeBtn, I remove the card in question, and then when adding
books, I only add the last one in the array. To avoid having to remove
everything before re-adding.